requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
Brotli==1.0.9
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dash==2.6.0
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dimod==0.11.6
diskcache==5.4.0
Django==4.1.3
djangorestframework==3.14.0
dwave-cloud-client==0.10.2
dwave-drivers==0.4.4
dwave-greedy==0.2.5
dwave-hybrid==0.6.9
dwave-inspector==0.3.0
dwave-inspectorapp==0.3.0
dwave-neal==0.5.9
dwave-networkx==0.8.12
dwave-ocean-sdk==5.5.0
dwave-preprocessing==0.4.0
dwave-system==1.15.0
dwave-tabu==0.4.5
dwavebinarycsp==0.2.0
fasteners==0.18
Flask==2.2.2
Flask-Compress==1.13
homebase==1.0.1
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
minorminer==0.2.9
mypy-extensions==0.4.3
networkx==2.8.8
numpy==1.23.4
packaging==21.3
pandas==1.3.5
parameterized==0.8.1
pathspec==0.10.1
penaltymodel==1.0.2
platformdirs==2.5.2
plotly==5.11.0
plucky==0.4.3
pluggy==1.0.0
psycopg2==2.9.5
py==1.11.0
pydantic==1.10.2
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
pytest-mock==3.10.0
python-dateutil==2.8.2
pytz==2022.6
rectangle-packer==2.0.1
requests==2.28.1
scipy==1.9.3
six==1.16.0
sqlparse==0.4.3
tenacity==8.1.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0
